Add targetless verification (= new default)
Verification is now intended to be sound for all supported targets (IP16, I16, ILP32, LLP64, and LP64) by default.

To get the old behavior, use '-target 32bit'.

Also added a Target architecture submenu to VFIDE's Verify menu.

Fixes #143. Also fixes #168.

+++ examples/stringBuffers.c
@@ -219,7 +219,7 @@ int chars_index_of_string(char *chars, int length, char *string)
             int cmp = memcmp(p, string, (size_t) n);
             //@ chars_join(p);
             //@ chars_join(chars);
-            if (cmp == 0) return p - chars;
+            if (cmp == 0) return (int)(p - chars);
             p++;
             //@ open string(string, stringChars);
             //@ chars_split(chars, p - chars);
